ci: move the module C under src/ and repoint every consumer
Step 12 of the skeleton adoption, and the one this module most needed: with
no src/, lint-c.sh, lint-nginx.sh, the gcovr filter and the CodeQL TU filter
all select nothing and PASS. That is the prompt's named empty-selection class,
not a cosmetic layout preference.

  filter/ngx_http_zstd_filter_module.c -> src/
  static/ngx_http_zstd_static_module.c -> src/
  ngx_http_zstd_common.h               -> src/
  ngx_http_zstd_sha256.h               -> src/

filter/config and static/config stay where they are: each module keeps its
own declaration fragment. The root config now points ngx_addon_dir at src/
before sourcing each, so both fragments resolve their sources against it and
their ../ climbs to the shared headers drop out entirely.

Same commit as the move, because nginx's config is the one file whose
breakage stops the module building at all.

Consumers repointed:
- src/*.c: the shared headers are siblings now, so "../ngx_http_zstd_*.h"
  becomes a plain "ngx_http_zstd_*.h".
- ci/fuzz/extract_parser.sh: slices the parser out of ../../src/.
- ci/tools/test_sha256_unit.c: includes ../../src/ngx_http_zstd_sha256.h.
- 12 clang-tidy/flawfinder/cppcheck invocations across build-test.yml,
  security-scanners.yml and ci-deep.yml that named the old paths by hand.
- .pre-commit-config.yaml's four C selectors collapse to
  '^(src/.*\.[ch]|ci/tools/.*\.[ch])$'.
- ci/fuzz/README.md's link and three stale path comments.

Verified, not assumed:
- NO_CACHE=1 ci/tools/ci-build.sh nginx 1.30.3 -- clean from-scratch rebuild
  succeeds, both .so produced.
- git ls-files 'src/*.[ch]' lists all four module sources and nothing else.
- prove ci/t/ -- 1654 tests pass (1195 + 459 static), unchanged from step 11.
- ci/fuzz/build.sh extracts from src/ and links.
- ci/tools/test_sha256_unit.sh passes, compiling the real header.
- the pre-commit C selector matches 10 tracked files including all four
  module sources -- a non-zero selection, checked rather than assumed.
- actionlint and shellcheck clean over everything touched.

DEFERRED, and it gates PR2: immediately after step 27 retargets the linter
checkers, a malloc/strcpy probe planted beside the module's real C must make
LINT_ONLY="c nginx" ci/linter/run-all.sh exit 1. A green there means the
selection is still empty and this move bought nothing.
