+++ ci/fuzz/fuzz_accept_encoding.c
@@ -12,7 +12,7 @@
  * misses, and it is a hard regression gate against re-introducing a
  * NUL-bounded scan over the len-bounded buffer.
  *
- * The real functions live in ../ngx_http_zstd_common.h and depend on the
+ * The real functions live in ../../src/ngx_http_zstd_common.h and depend on the
  * nginx tree. fuzz/extract_parser.sh slices just those functions into
  * generated_parser.inc at build time, so we always fuzz the SHIPPED code
  * with no copy drift. ngx_shim.h provides the tiny nginx surface it needs
@@ -30,7 +30,7 @@
 
 #include "ngx_shim.h"
 
-/* Generated by fuzz/extract_parser.sh from ../ngx_http_zstd_common.h.
+/* Generated by ci/fuzz/extract_parser.sh from ../../src/ngx_http_zstd_common.h.
  * Contains the verbatim bodies of ngx_http_zstd_eval_qvalue() and
  * ngx_http_zstd_accept_encoding(). */
 #include "generated_parser.inc"
