+++ filter/config
@@ -10,8 +10,8 @@ HTTP_ZSTD_SRCS="$ngx_addon_dir/ngx_http_zstd_filter_module.c"
 # dependencies so an incremental build (nginx's auto/sources / make .d
 # generation) rebuilds the object whenever a header changes; without
 # this, edits to them leave stale .o files behind.
-HTTP_ZSTD_DEPS="$ngx_addon_dir/../ngx_http_zstd_common.h \
-                $ngx_addon_dir/../ngx_http_zstd_sha256.h"
+HTTP_ZSTD_DEPS="$ngx_addon_dir/ngx_http_zstd_common.h \
+                $ngx_addon_dir/ngx_http_zstd_sha256.h"
 
 ngx_addon_name=ngx_http_zstd_filter_module
 ngx_module_type=HTTP_FILTER
