+++ static/config
@@ -11,7 +11,7 @@ HTTP_ZSTD_SRCS="$ngx_addon_dir/ngx_http_zstd_static_module.c"
 # dependency so an incremental build rebuilds the object whenever the
 # header changes; without this, edits to ngx_http_zstd_common.h leave
 # stale .o files behind.
-HTTP_ZSTD_DEPS="$ngx_addon_dir/../ngx_http_zstd_common.h"
+HTTP_ZSTD_DEPS="$ngx_addon_dir/ngx_http_zstd_common.h"
 
 ngx_addon_name=ngx_http_zstd_static_module
 ngx_module_type=HTTP
